feat: add buttons to add and subtract numbers.

+++ src/SS.Core/IO/Files/Settings/SGameplaySettings.cs
@@ -12,13 +12,13 @@ namespace StardustSandbox.Core.IO.Files.Settings
         [XmlElement("PreviewAreaColor", typeof(Color))]
         public Color PreviewAreaColor { get; set; }
 
-        [XmlElement("PreviewAreaOpacity", typeof(float))]
-        public float PreviewAreaOpacity { get; set; }
+        [XmlElement("PreviewAreaOpacity", typeof(byte))]
+        public byte PreviewAreaOpacity { get; set; }
 
         public SGameplaySettings()
         {
             this.PreviewAreaColor = SColorPalette.White;
-            this.PreviewAreaOpacity = 25f;
+            this.PreviewAreaOpacity = 25;
         }
     }
 }
